+++ WhistleWind.AbnormalSigils/Abilities/abilityPiercing.cs
@@ -17,7 +17,7 @@ namespace WhistleWind.AbnormalSigils
 
             Piercing.ability = AbnormalAbilityHelper.CreateAbility<Piercing>(
                 Artwork.sigilPiercing, Artwork.sigilPiercing_pixel,
-                rulebookName, rulebookDescription, dialogue, powerLevel: 2,
+                rulebookName, rulebookDescription, dialogue, powerLevel: 1,
                 modular: true, opponent: false, canStack: false).Id;
         }
     }